Fix wrong values in certain categories/months.
The reason was, that when I joined the transactions and splittransactions, I selected for every splittransaction the amount of the original transaction instead of the amount of the splittransaction.
Example: Splittransaction 9 = 3 + 4 + 2 became 3 transaction, with the value of each = 9, instead of 3 transaction where one has the value 3, one the value 4 and one the value 2.